Remove userDefaults, change class name
